Moved ManHuntRuleCommandExecutor to settings-package
